Changes that hadn't been persisted from before the last batch of changes?

+++ jbombardier-agent/src/main/java/com/jbombardier/agent/Agent2.java
@@ -696,7 +696,7 @@ public class Agent2 implements Asynchronous {
                     threadController.setRateStepTime((Long) newValue);
                     break;
                 case targetRate:
-                    threadController.setTargetRate((Float) newValue);
+                    threadController.setTargetRate(((Double)newValue).floatValue());
                     break;
                 case targetThreads:
                     threadController.setTargetThreads((Integer) newValue);
